requirements.txt
# Core dependencies
django>=4.2.0
djangorestframework>=3.14.0
Pillow>=10.0.0
numpy>=1.24.0
opencv-python>=4.8.0
torch>=2.0.0
torchvision>=0.15.0
faiss-cpu>=1.7.4
scikit-learn>=1.3.0

# Image processing and ML
sentence-transformers>=2.2.2
easyocr>=1.7.0
pytesseract>=0.3.10
pyzbar>=0.1.9

# Optional but recommended
python-dotenv>=1.0.0
redis>=5.0.0
psycopg2-binary>=2.9.0  # For PostgreSQL support 